% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_daily_top_ten_tbl}
\alias{create_daily_top_ten_tbl}
\title{Title}
\usage{
create_daily_top_ten_tbl(
  token = "FLIX_PATROL",
  platform_name,
  country_name,
  date,
  flix_type
)
}
\arguments{
\item{flix_type}{}
}
\description{
Title
}
