% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_hours_viewed_tbl}
\alias{create_hours_viewed_tbl}
\title{Title}
\usage{
create_hours_viewed_tbl(
  token = "FLIX_PATROL",
  platform_name = "netflix",
  media_type_name,
  language_name,
  start_date,
  end_date
)
}
\arguments{
\item{language_name}{}
}
\value{
tibble
}
\description{
Title
}
