% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{get_title_name}
\alias{get_title_name}
\title{Get Title Name from ID}
\usage{
get_title_name(fp_id, token = "FLIX_PATROL")
}
\arguments{
\item{token}{Character. API environment variable name.}

\item{id_prefix}{Character. The FlixPatrol title ID (e.g., "ttl_bHyGTvopBHPVtIKhR2CF68WD").}
}
\value{
A character string containing the title name.
}
\description{
Retrieves the official movie or TV show title associated with a FlixPatrol ID.
}
\details{
Accesses the specific resource path `/titles/:id`. Useful for cleaning
up datasets where only IDs are present.
}
