% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lookup_country}
\alias{lookup_country}
\title{Lookup Country IDs}
\usage{
lookup_country(country_name, silent = getOption("flixpatrol.silent", FALSE))
}
\arguments{
\item{country_name}{Character vector. The name(s) of countries to lookup (e.g., "United States", "France").}

\item{silent}{Logical. If `FALSE` (default), prints success messages using `cli`.}
}
\value{
A named character vector of country IDs.
}
\description{
Translates country names into FlixPatrol internal country IDs.
}
\details{
This function queries the `/countries` endpoint using exact matching. It is
vectorized via `purrr::map_chr`, meaning you can pass a single country name
or a vector of names. It will return the matching ID for each name provided.
}
