% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_country_and_return_id}
\alias{lookup_country_and_return_id}
\title{Title}
\usage{
lookup_country_and_return_id(
  country_name,
  token = "FLIX_PATROL",
  silent = FALSE
)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
