% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lookup_date_type}
\alias{lookup_date_type}
\title{Lookup Date Type IDs}
\usage{
lookup_date_type(x)
}
\arguments{
\item{x}{Character vector. Date types (e.g., "week") or `"*"` for all.}
}
\value{
An integer vector of IDs.
}
\description{
Resolves date types (Day, Week, Month) to internal API IDs.
}
