% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_date_type_name_and_return_id}
\alias{lookup_date_type_name_and_return_id}
\title{Lookup Date Type IDs}
\usage{
lookup_date_type_name_and_return_id(x)
}
\arguments{
\item{x}{Character vector. Date types (e.g., "week").}
}
\value{
An integer vector of IDs.
}
\description{
Resolves date types (Day, Week, Month) to internal API IDs.
}
