% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{lookup_language}
\alias{lookup_language}
\title{Lookup Language IDs}
\usage{
lookup_language(language_name)
}
\arguments{
\item{language_name}{Character vector. Language names or `"*"` for all.}
}
\value{
An integer vector of IDs.
}
\description{
Resolves language names to internal FlixPatrol IDs.
}
