% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_language_name_and_return_id}
\alias{lookup_language_name_and_return_id}
\title{Lookup Language IDs}
\usage{
lookup_language_name_and_return_id(language_name)
}
\arguments{
\item{language_name}{Character vector. Language names or "*" for all.}
}
\value{
An integer vector of IDs.
}
\description{
Resolves language names to internal FlixPatrol IDs.
}
