% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_media_type}
\alias{lookup_media_type}
\title{Lookup Media Type IDs}
\usage{
lookup_media_type(media_type_name)
}
\arguments{
\item{media_type_name}{Character vector. Media types or `"*"` for all.}
}
\value{
An integer vector of IDs.
}
\description{
Resolves media type labels to internal FlixPatrol IDs.
}
