% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_media_type_name_and_return_id}
\alias{lookup_media_type_name_and_return_id}
\title{Title}
\usage{
lookup_media_type_name_and_return_id(media_type_name)
}
\arguments{
\item{media_type_name}{}
}
\description{
Title
}
