% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_platform_and_return_id}
\alias{lookup_platform_and_return_id}
\title{Title}
\usage{
lookup_platform_and_return_id(
  platform_name,
  token = "FLIX_PATROL",
  silent = FALSE
)
}
\arguments{
\item{silent}{}
}
\description{
Title
}
