% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{lookup_title}
\alias{lookup_title}
\title{Lookup Title ID}
\usage{
lookup_title(title, silent = getOption("flixpatrol.silent", FALSE))
}
\arguments{
\item{title}{Character vector. The name of the movie or TV show.}

\item{silent}{Logical. If FALSE, prints success messages.}
}
\value{
A named character vector of title IDs.
}
\description{
Translates movie or TV show titles into FlixPatrol internal title IDs.
}
\details{
Queries the `/titles` endpoint using exact matching. This is vectorized,
allowing for multiple title lookups at once.
}
