% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_torrent_site}
\alias{lookup_torrent_site}
\title{Lookup Torrent Site IDs}
\usage{
lookup_torrent_site(x)
}
\arguments{
\item{x}{Character vector. Torrent site names or `"*"` for all.}
}
\value{
A character vector of IDs.
}
\description{
Resolves torrent site names to internal FlixPatrol IDs.
}
