% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_country_and_return_id}
\alias{lookup_country_and_return_id}
\title{Lookup Country IDs}
\usage{
lookup_country_and_return_id(
  country_name,
  token = "FLIX_PATROL",
  silent = FALSE
)
}
\arguments{
\item{country_name}{Character vector. The name(s) of countries to lookup (e.g., "USA", "France").}

\item{token}{Character. The environment variable name for the API key.}

\item{silent}{Logical. If `FALSE` (default), prints success messages using `cli`.}
}
\value{
A named character vector of country IDs.
}
\description{
Translates country names into FlixPatrol internal country IDs.
}
\details{
This function queries the `/countries` endpoint using a "like" search. It is
vectorized via `purrr::map_chr`, meaning you can pass a single country name
or a vector of names. It will return the first matching ID for each name provided.
}
