% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lookup_language_name_and_return_id}
\alias{lookup_language_name_and_return_id}
\title{Title}
\usage{
lookup_language_name_and_return_id(language_name)
}
\arguments{
\item{language_name}{}
}
\description{
Title
}
