% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{validate_date_range}
\alias{validate_date_range}
\title{Title}
\usage{
validate_date_range(
  start_date,
  end_date,
  start_date_wday_abb,
  range_length = 7
)
}
\arguments{
\item{range_length}{}
}
\value{
logical
}
\description{
Title
}
